+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [v1.2.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.1.0...v1.1.0) - 2021-10-11
+### Added
+- `toExistInStorage()` expectation
+
+### Fixed
+- [Pest IntelliJ](https://github.com/pestphp/pest-intellij) support for autocompletion
+
 ## [v1.1.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.0.4...v1.1.0) - 2021-10-08
 ### Added
 - `toBeAbleTo()` expectation